personalProjects/Translator/requirements.txt
flask
python-dotenv
requests
azure-cognitiveservices-speech
random